+++ src/termout.c
@@ -914,6 +914,26 @@ could_be_emoji_base(termchar * tc)
          );
 }
 
+/*
+   Determine characters for Arabic Lam/Alef single-cell joining
+   Unicode has presentation forms (isolated and final) only for 
+	U+644 LAM
+   combined with either of
+	U+627 ALEF
+	U+622 ALEF WITH MADDA ABOVE
+	U+623 ALEF WITH HAMZA ABOVE
+	U+625 ALEF WITH HAMZA BELOW
+	U+649 ALEF MAKSURA
+   (where the ligature with ALEF MAKSURA, however, is wider than one cell),
+   but not for any other LAM WITH ... (SMALL V, DOT ABOVE, etc) or
+   any other ALEF (WASLA, WITH WAVY HAMZA, etc);
+   so, lacking more information about Arabic typography, the 
+   assumption is that only the combinations of plain LAM with 
+   ALEF or ALEF WITH MADDA or WITH HAMZA ABOVE or BELOW need to be supported
+ */
+bool isLAM(xchar c) { return c == 0x644; }
+bool isALEF(xchar c) { return c >= 0x622 && c <= 0x627 && c != 0x624 && c != 0x626; }
+
 static wchar last_high = 0;
 static wchar last_char = 0;
 static int last_width = 0;
@@ -974,6 +994,16 @@ write_char(wchar c, int width)
     //TODO: if changed, propagate mode onto paragraph
     if (cfg.ligatures_support)
       term_invalidate(0, curs->y, curs->x, curs->y);
+  }
+
+  // check for Arabic Lam/Alef single-cell joining
+  if (term.join_lam_alef &&
+      curs->x && isALEF(c) && isLAM(line->chars[curs->x - 1].chr)
+     )
+  {
+    // in LAM/ALEF single-cell joining mode, handle ALEF after LAM like a 
+    // combining character, in order to trigger their single-cell rendering
+    width = 0;
   }
 
   if (curs->wrapnext && term.autowrap && width > 0) {
@@ -2527,6 +2557,8 @@ set_modes(bool state)
             term.curs.bidimode &= ~LATTR_BIDISEL;
           else
             term.curs.bidimode |= LATTR_BIDISEL;
+        when 2521:      /* LAM/ALEF single-cell joining */
+          term.join_lam_alef = state;
         when 2026:
           term.suspend_update = state ? 150 : 0;
           if (!state) {
@@ -2704,6 +2736,8 @@ get_mode(bool privatemode, int arg)
         return 2 - !!(term.curs.bidimode & LATTR_BOXMIRROR);
       when 2501: /* bidi direction auto-detection */
         return 2 - !(term.curs.bidimode & LATTR_BIDISEL);
+      when 2521: /* LAM/ALEF single-cell joining */
+        return 2 - term.join_lam_alef;
       when 7723: /* Reflow mode; 2027 is dropped */
         return 2 - term.curs.rewrap_on_resize;
       when 2027 or 7769: /* Emoji 2-cell width mode */