+++ src/termpriv.h
@@ -39,6 +39,10 @@ extern void clear_wrapcontd(termline * line, int y);
 extern ushort getparabidi(termline * line);
 extern wchar * wcsline(termline * line);  // for debug output
 
+/* Advanced Arabic support */
+extern bool isLAM(xchar c);
+extern bool isALEF(xchar c);
+
 static inline bool
 term_selecting(void)
 { return term.mouse_state < 0 && term.mouse_state >= MS_SEL_LINE; }